Recator annotation export and run on annotation change
